Add data and script from M2 version + classifies files according to M1/M2 versions
Here we add the script and data used for the same project in M2. Also we create a folder for the fisrt version done in M1 and in the 2021 data challenge ; and a folder for the M2 version. Both are not using the same datas, but it is still the same topic.